requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
emoji==1.6.3
emojis==0.6.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pillow==9.0.1
pip==21.2.4
pytz==2021.3
setuptools==58.1.0
telegram==0.0.1
yarl==1.7.2